+++ examples/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.9.26
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.156
+  - antd-management-fast-design-react@1.0.95
+  - antd-management-fast-component@2.7.13
+  - antd-management-fast-framework@2.12.13
+  - antd-management-fast-formily@1.0.105
+  - antd-management-fast-common@2.6.126
+  - antd-management-fast-flow@1.0.95
+
 ## 1.9.25
 
 ### Patch Changes